+++ CHANGES.rst
@@ -7,6 +7,9 @@ Unreleased
 
 * Bump ``sql_exporter`` to ``0.18.0``.
 
+* Always set ``app.kubernetes.io/managed-by`` to ``crate-operator`` to ensure the
+  operator can reliably identify CrateDB pods.
+
 2.48.0 (2025-06-10)
 -------------------